refactor: streamline TiptapEditor and introduce new hooks for improved functionality
- Refactor TiptapEditor component by removing unused imports and consolidating state management.
- Introduce new hooks: useEditorLifecycle, useEditorSetup, useSuggestionEffects, and useThumbnailCommit for better separation of concerns.
- Update collaboration handling in PageEditorLayout and ensure proper state dependencies in usePageEditorEffects.
- Enhance suggestion handling and thumbnail insertion logic for improved user experience.
