+++ commitlint.config.js
@@ -1,1 +1,6 @@
-export default { extends: ["@commitlint/config-conventional"] };
+export default {
+  extends: ["@commitlint/config-conventional"],
+  rules: {
+    "body-max-line-length": [0],
+  },
+};
